+++ packages/keyring-internal-snap-client/src/index.ts
@@ -1,2 +1,1 @@
 export * from './KeyringInternalSnapClient';
-export * from './v2';