+++ packages/keyring-snap-sdk/src/index.ts
@@ -2,4 +2,3 @@ export * from './rpc-handler';
 export * from './snap-utils';
 export * from './time';
 export * from './methods';
-export * from './v2';