+++ packages/keyring-eth-hd/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `./v2` subpath export for keyring v2 implementation ([#513](https://github.com/MetaMask/accounts/pull/513))
+  - `HdKeyring` and `HdKeyringOptions` are now available from `@metamask/eth-hd-keyring/v2`.
+
+### Changed
+
+- **BREAKING:** Move and rename `HdKeyringV2` and `HdKeyringV2Options` to the new `./v2` subpath export ([#513](https://github.com/MetaMask/accounts/pull/513))
+  - `HdKeyringV2` is now `HdKeyring` from `@metamask/eth-hd-keyring/v2`.
+  - `HdKeyringV2Options` is now `HdKeyringOptions` from `@metamask/eth-hd-keyring/v2`.
+
 ## [13.1.1]
 
 ### Changed